+++ sdk/core/azure-core/src/main/java/com/azure/core/implementation/util/FluxByteBufferContent.java
@@ -3,6 +3,7 @@
 
 package com.azure.core.implementation.util;
 
+import com.azure.core.implementation.FluxInputStream;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.serializer.ObjectSerializer;
@@ -102,7 +103,7 @@ public final class FluxByteBufferContent extends BinaryDataContent {
 
     @Override
     public InputStream toStream() {
-        return new ByteArrayInputStream(toBytes());
+        return isReplayable ? new ByteArrayInputStream(toBytes()) : new FluxInputStream(content);
     }
 
     @Override